requirements.txt
streamlit>=1.32,<2.0
pandas>=2.1
numpy>=1.26
plotly>=5.15
python-dateutil>=2.8
openpyxl>=3.1
XlsxWriter>=3.1
reportlab>=4.0
pyarrow
duckdb
pmdarima
sqlmodel
duckdb-engine
pyyaml
openai
requests
statsmodels
python-pptx
scipy
scikit-learn
chardet
ruff
black
pytest
hypothesis
transformers>=4.36
torch
